Fixed hand animations.
Input is ready for play testing and tweaking.  If necessary, the player can use the left joystick to move and the right joystick to turn.  However, this is only necessary if the real play zone is smaller than the virtual play zone.

Playing with the purple block is a bit glitchy, however, as its mass can launch the player in the air.