Create Settings.cs
Prototype for how settings can be handled.  This stores values changed in the ui, saves and loads the settings so that the user does not need to set them every time they boot the simulation or change scene, and *possibly, this needs testing* manipulates the values that they should manipulate.  To use this script, add it to a game object (can be empty except for this script) and then drag it to the list of actions that are triggered by the button in the settings menu that closes that menu. Then select SaveChanges as the function to call.