Prototype kinda works
The "links" are staying together, but they are extremely jittery.  A new prototype may rely on a script to spawn and connect the links so that they can be made very small and uniform.  It would probably involve a for loop in a method that takes in the number of links desired and uses that as the number as the iteration limiter.